add purchase lines with float value

+++ src/actions/CRUD/purchLine/createPurchLine.ts
@@ -11,7 +11,7 @@ export async function createPurchLine(
             console.log('createPurchLineRecord', formData)
             const purchase_id: number = parseInt(formData.get('purchase_id') as string);
             const item_id: number = parseInt(formData.get('item_id') as string);
-            const quantity: number = parseInt(formData.get('quantity') as string);
+            const quantity: number = parseFloat(formData.get('quantity') as string);
             const unit_id: number = parseInt(formData.get('unit_id') as string);
                         
             await db.purchaselines.create({
